requirements.txt
aiogram
requests
openai
aiohttp
python-dotenv
vosk
aiofiles
sqlalchemy
psycopg2
rutimeparser